requirements.txt
gunicorn==21.2.0
django==4.2.10
pip==23.3.2
django-allauth==0.56.0
whitenoise
django-heroku==0.3.1